+++ po-includes/database/migrations/2019_10_10_081154_create_comments_table.php
@@ -14,16 +14,16 @@ class CreateCommentsTable extends Migration
     public function up()
     {
         Schema::create('comments', function (Blueprint $table) {
-            $table->increments('id');
-			$table->integer('parent')->default('0');
-			$table->integer('post_id')->nullable();
+            $table->bigIncrements('id');
+			$table->bigInteger('parent')->default('0');
+			$table->bigInteger('post_id')->nullable();
 			$table->string('name')->nullable();
 			$table->string('email')->nullable();
 			$table->text('content')->nullable();
 			$table->enum('active', ['Y', 'N'])->default('N');
 			$table->enum('status', ['Y', 'N'])->default('N');
-			$table->integer('created_by')->default('1');
-			$table->integer('updated_by')->default('1');
+			$table->bigInteger('created_by')->default('1');
+			$table->bigInteger('updated_by')->default('1');
             $table->timestamps();
         });
     }
